// src/store/flow/mutation-types.ts
export enum Types {
  CHANGE_SECTION_STYLE = "CHANGE_SECTION_STYLE",
  INSERT_TO_BOTTOM = "INSERT_TO_BOTTOM",
  CHANGE_INDEX = "CHANGE_INDEX",
  CHANGE_SECTION = "CHANGE_SECTION",
  CHANGE_SECTION_HEIGHT = "CHANGE_SECTION_HEIGHT",
  AUTO_ADJUST_SECTION_HEIGHT = "AUTO_ADJUST_SECTION_HEIGHT",
  ADD_NEW_SECTION_IN_FRONT_CURRENT = "ADD_NEW_SECTION_IN_FRONT_CURRENT",
  ADD_NEW_SECTION_AT_BACK_CURRENT = "ADD_NEW_SECTION_AT_CURRENT_BACK",
  MOVE_SECTION_FORWARD = "MOVE_SECTION_FORWARD",
  MOVE_SECTION_BACKWARD = "MOVE_SECTION_BACKWARD",
  CHANGE_BACKGROUND_HIGHLIGHT_INDEX = "CHANGE_BACKGROUND_HIGHLIGHT_INDEX",
  REMOVE_SECTION = "REMOVE_SECTION",
  CHANGE_COMPONENT = "CHANGE_COMPONENT",
  CHANGE_COMPONENT_LEFT = "CHANGE_COMPONENT_LEFT",
  CHANGE_COMPONENT_TOP = "CHANGE_COMPONENT_TOP",
  CHANGE_COMPONENT_STYLE = "CHANGE_COMPONENT_STYLE",
  CHANGE_OPERATOR_STYLE = "CHANGE_OPERATOR_STYLE",
  CHANGE_OPERATOR_HEIGHT = "CHANGE_OPERATOR_HEIGHT",
  CHANGE_IS_DRAG_START = "CHANGE_IS_DRAG_START",
  CHANGE_IS_DRAG_DROP = "CHANGE_IS_DRAG_DROP",
  ADD_TEMP_SECTION = "ADD_TEMP_SECTION",
  REMOVE_TEMP_SECTION = "REMOVE_TEMP_SECTION",
  CHANGE_DRAG_HIGHLIGHT_INDEX = "CHANGE_DRAG_HIGHLIGHT_INDEX",
  ADD_NEW_COMPONENT_TO_NEW_SECTION = "ADD_NEW_COMPONENT_TO_NEW_SECTION",
  ADD_NEW_COMPONENT_TO_EXIST_SECTION = "ADD_NEW_COMPONENT_TO_EXIST_SECTION",
}